backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
httpx==0.27.0
groq==0.9.0
gitpython==3.1.43
sentence-transformers==2.7.0
langchain==0.2.6
langchain-groq==0.1.6
langgraph==0.1.5
spacy==3.7.4
PyGithub==2.3.0
pydantic==2.7.1
python-multipart==0.0.9
